Fix up testing for Puppet 4
With the release of Puppet 4, some things in the internals or Puppet may be
changing. This travis-ci configuration did not have an awareness of Puppet 4.

This commit adds awareness and testing for the latest Puppet 3 and latest
Puppet 4 series releases. This commit also removes testing for the Puppet 2.7
series which means that Puppet 2.7 is no longer supported by this module.