Use node facts and trused data for Console search
This commit extends the noop_from_console function to use node facts and
trausted data when doing the lookup of classification. This will result
in actual results being returned based on rule matching and not just
node pinning.

This commit also adds some simple improvements to the function based on
the results of manual testing.