Add expect to basic compile test
Without this commit there was a test in the unit testing suite that did
not have any expectations around the outcome of the actions. Now that we
are on a newer Rspec, expectations should be used as part of the tests
fairly consistently. The test in question was simply ensuring that the
function could compile a catalog with zero arguments were passed.
Possibly a useless test since it's ability to work is verified in other
tests. Nonetheless it should be written correctly.

This commit adds the missing expect statement.