Rubocop style corrections
With the addition of PDK to the development process, evaluation of the
style of the Ruby code is not part of the standard checks. The tooling
that performs the evaluation, Rubocop, has identified a number of ways
the increase the readability and maintainability of the Ruby code.

This commit applies the recommendations from Rubocop.