Rewrite of unit tests
Without this commit, the unit tests included in this module were
becoming stale and outdated. As the function(s) in this module have been
updated to take advantage of the new Puppet 4 function API, the testing
suite was no longer providing any value.

This commit adds the refactored tests using the latest features and
versions of rspec-puppet. All tests should pass and correctly validate
all behaviors of the noop module.