Fix a bug with resource collector overrides causing compilation failure
When resource collectors were used to override parameters in a child
scope of the noop() call compilation would fail with the message
"Parameter 'noop' is already set on <resource>; cannot redefine".

This is because the resource collector internally creates a new resource
for overrides (which gets a noop parameter) and merges it with the
collected resource (which already has a noop parameter). This triggers
parameter override logic.

The parameter override logic checks to see if:
  (1) the new parameter source is a child of the original parameter source
 or if:
  (2) the parameter is a default that is being overridden.

The resource collector normally hijacks the code to say that it is a child
of the original source (fulfilling condition (1) ), but this doesn't
propagate to the noop parameter that gets added. Thus the noop parameter
doesn't match either side of the conditional and fails.

This commit patches this logic by adding the noop parameter as a default
on every resource type that it touches. This allows it to match
condition (2) and allows compilation to continue.

See also:
  override_parameter in lib/puppet/parser/resource.rb
  evaluate in lib/puppet/pops/evaluator/collectors/abstract_collector.rb