Add noop(undef) capability, to reset noop default
The noop() function is powerful because it persists a default setting
through nested scope layers. Sometimes, it is useful to set an outer
scope default to one value (say, true), but set a small inner scope
default to another value (such as the already supported false).

This commit adds the ability to un-set a noop default at an inner-scope
layer. This is useful when a user wants to do something similar to
noop(false) inside an outer scope that has had noop(true) called, but
wishes to preserve the full behavior of `--noop` on the command line. In
that case, noop(false) is insufficient because that would cause `--noop`
to be largely ignored; individual resource-level noop parameters take
precedence over a CLI --noop argument.

The new noop(undef) solves for this problem. It will remove any
enclosing scope's noop_default, but not affect the behavior of CLI
--noop.